pyproject.toml
[tool.poetry]
name = "testtask"
version = "0.1.0"
description = ""
authors = ["ramazan_aisaev <aisaev-68@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.9.1"
fastapi = "^0.103.2"
asyncpg = "^0.28.0"
pydantic = "^2.4.2"
aiofiles = "^23.2.1"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
types-aiofiles = "^23.2.0.0"
pandas = "1.5.0"
pydantic-settings = "2.0.2"
jinja2 = "^3.1.2"
openpyxl = "^3.1.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.21"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
